feat: Add order cancellation and stock disposal features
- Implemented cancelOrder functionality in IOrderUseCasePort and OrderUseCase.
- Added CancelOrderCommand and CancelOrderResultDto for handling order cancellation requests.
- Created CancelOrderRequest and CancelOrderResponse DTOs for API interaction.
- Introduced DisposeLot functionality in IStockInventoryUseCasePort and StockInventoryUseCase.
- Added DisposeLotCommand and DisposeLotResultDto for managing stock disposal operations.
- Created DisposeLotRequest and DisposeLotResponse DTOs for API interaction.
- Added ProductStatsDto and ProductStatsResponse for retrieving product and inventory statistics.
- Updated IProductUseCasePort and ProductUseCase to include getProductStats method.
- Enhanced IProductRepository and IStockInventoryRepository with new statistics methods.
- Updated ProductDbo to change totalStockQuantity type from Integer to Long for better handling of stock quantities.
- Added necessary mappings and queries in ProductAdapter and StockInventoryAdapter.
- Updated controllers to handle new endpoints for cancelling orders and disposing lots.

+++ src/main/java/com/fivetpromart/infrastructure/persistence/product/ProductDbo.java
@@ -34,7 +34,7 @@ public class ProductDbo {
     BigDecimal sellingPrice;
 
     @Column(name = "total_stock_quantity")
-    private Integer totalStockQuantity;
+    private Long totalStockQuantity;
 
 //    @Transient
 //    @Builder.Default
